[DatePickerHeader]: moved to uui, reworked buttons, [DatePickerBody]: moved to uui, removed redundant code, fixed styles, [RangeDatePickerBody]: moved to uui, removed redundant code, fixed styles, [Filters]: fixed empty values in toggler for datepicker/rangedatepicker updated themes
