+++ uui-components/src/inputs/index.ts
@@ -6,15 +6,14 @@ export * from './Rating/Rating';
 export * from './Rating/BaseRating';
 export * from './NumericInput';
 export * from './TimePickerBody';
-export * from './DatePicker/DatePickerBody';
-export * from './DatePicker/DatePickerHeader';
 export * from './DatePicker/Day';
 export * from './DatePicker/Calendar';
 export * from './DatePicker/CalendarPresets';
-export * from './DatePicker/RangeDatePickerBody';
 export * from './DatePicker/DatePickerBodyBase';
 export * from './DatePicker/BaseDatePicker';
 export * from './DatePicker/BaseRangeDatePicker';
+export * from './DatePicker/MonthSelection';
+export * from './DatePicker/YearSelection';
 export * from './DatePicker/helpers';
 export * from './Slider/SliderBase';
 export * from './Slider/RangeSlider';
