+++ uui/components/datePickers/RangeDatePickerBody.module.scss
@@ -87,3 +87,39 @@
         border-left: 1px solid var(--uui-control-border);
     }
 }
+
+.container {
+    &.day-selection {
+        .from-picker {
+            :global(.uui-datepickerheader-nav-icon-right) {
+                visibility: hidden;
+            }
+        }
+
+        .to-picker {
+            :global(.uui-datepickerheader-nav-icon-left) {
+                visibility: hidden;
+            }
+        }
+    }
+
+    .bodes-wrapper {
+        position: relative;
+        transition: height 0.3s ease-in-out;
+    }
+
+    .blocker {
+        background-color: rgba(255, 255, 255, 0.9);
+        position: absolute;
+        left: 0;
+        right: 0;
+        top: 0;
+        bottom: 0;
+    }
+}
+
+:global(.uui-datepicker-container) {
+    transition: height 0.3s ease-in-out;
+}
+
+
