+++ uui/components/datePickers/__tests__/CalendarPresets.test.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import * as React from 'react';
 import { CalendarPresets } from '../CalendarPresets';
 import { renderer } from '@epam/uui-test-utils';
-import { rangeDatePickerPresets } from '@epam/uui-components';
+import { rangeDatePickerPresets } from '../RangeDatePickerBody';
 
 describe('CalendarPresets', () => {
     it('should be rendered correctly', () => {
