+++ uui/components/datePickers/__tests__/DatePickerHeader.test.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+// @ts-ignore
 import dayjs from 'dayjs';
 import { fireEvent, setupComponentForTest, screen, within } from '@epam/uui-test-utils';
 import { DatePickerHeader, DatePickerHeaderProps } from '../DatePickerHeader';
