+++ uui/components/layout/FlexItems/FlexCell.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import css from './FlexRow.module.scss';
 import { FlexCellProps } from '@epam/uui-core';
 import { FlexCell as uuiFlexCell } from '@epam/uui-components';
@@ -6,4 +5,4 @@ import { withMods } from '@epam/uui-core';
 
 export type FlexCellMods = {};
 
-export const FlexCell = withMods<FlexCellProps, FlexCellMods>(uuiFlexCell, (props) => [css.flexCell]);
+export const FlexCell = withMods<FlexCellProps, FlexCellMods>(uuiFlexCell, () => [css.flexCell]);
