+++ uui/components/navigation/MainMenu/MainMenu.tsx
@@ -1,7 +1,8 @@
 import css from './MainMenu.module.scss';
 import { withMods } from '@epam/uui-core';
 import { MainMenu as uuiMainMenu, MainMenuProps } from '@epam/uui-components';
-import { Burger, MainMenuDropdown } from '../MainMenu';
+import { Burger } from './Burger';
+import { MainMenuDropdown } from './MainMenuDropdown';
 
 export interface MainMenuMods {}
