+++ uui/components/overlays/ConfirmationModal.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { IModal } from '@epam/uui-core';
-import {
-    ModalBlocker, ModalWindow, FlexSpacer, ModalHeader, Button, ModalFooter, Panel, ScrollBars,
-} from '../';
+import { Button } from '../buttons';
+import { FlexSpacer, Panel, ScrollBars } from '../layout';
+import { ModalBlocker, ModalWindow, ModalHeader, ModalFooter } from './Modals';
 import { i18n } from '../../i18n';
 
 export interface ConfirmationModalWindowProps extends IModal<any> {
